Add synthesize-set-identifier flag
@devkid added functionality to external-dns to allow multiple records with the
same FQDN but different targets, derived from different K8s resources.  This is
exactly what we need, but the implementation depends on the DNS provider
persisting some metadata -- a set identifier -- with each DNS record.  OCI DNS
and many other DNS providers do not support storing record-scoped metadata.

This change adds a boolean flag that triggers new behavior after the DNS
records and K8s endpoints are fetched, but before the plan is calculated.  The
endpoints are scanned for non-empty set-identifiers (picked up from annotations
on the ingress/service), and a map is built, keyed on the endpoint details.
Then the records are scanned, and when the endpoint details match the record
details, the associated set-identifier is set on the record.  This simulates
the behavior observed when the set-identifier is persisted with the DNS record,
and allows the plan calculation to retain multiple records with the same FQDN.